Stricter input parameter validation
Replicated exactly the way dinput.dll validates instance and version parameters.
This also fixes a potential issue where a version 0x300 DirectInput was created
with a null hInstance, and that was then passed to DirectInput8 - but unlike
in Di3, this is not permitted in Di8. In those cases we now pass a handle
to the main module.