Emulate old DInput behaviour of EnumObjects
EnumObjects of old DirectInput differs from DirectInput8 in three ways:
1. When the data format is set, objects not specified in that format get
   an offset of -1 (all DInput versions).
2. Objects are enumerated in order and with offsets that match a predefined data format,
   while DInput8 enumerates them in "raw data" order (all DInput versions except for 0x700 and 0x5B2).
3. Collections are not enumerated, except for versions 0x5B2 and 0x700

Enabled C++17 to get push_back to return a reference to the new element,
it's very handy.

Updated the "Supported Games" list with racing games from DICE that
prompted this research in the first place. As of this commit, they all work fine with dinputto8.