Various conformance improvements (#21)
* Better keep track of a requested DirectInput version

Made diVersion a class variable as there's technically nothing preventing
apps from creating multiple instances of DInput objects with different versions.
Fixed a few cases of the emulated version "leaking" into DInput8 calls,
probably causing failures.

* Stricter input parameter validation

Replicated exactly the way dinput.dll validates instance and version parameters.
This also fixes a potential issue where a version 0x300 DirectInput was created
with a null hInstance, and that was then passed to DirectInput8 - but unlike
in Di3, this is not permitted in Di8. In those cases we now pass a handle
to the main module.

* Do not enumerate game controllers for DirectInput 0x300 and older

This matches the original dinput.dll's behaviour.

* Slightly refactor a dwDevType == DI8DEVCLASS_ALL case in EnumDevicesX

- Less copying data around (the final list stores references)
- Less unused data by not reusing DeviceEnumerator for no benefit

* Use local variables in IDirectInputDeviceX::SetDataFormat

* Emulate old DInput behaviour of EnumObjects

EnumObjects of old DirectInput differs from DirectInput8 in three ways:
1. When the data format is set, objects not specified in that format get
   an offset of -1 (all DInput versions).
2. Objects are enumerated in order and with offsets that match a predefined data format,
   while DInput8 enumerates them in "raw data" order (all DInput versions except for 0x700 and 0x5B2).
3. Collections are not enumerated, except for versions 0x5B2 and 0x700

Enabled C++17 to get push_back to return a reference to the new element,
it's very handy.

Updated the "Supported Games" list with racing games from DICE that
prompted this research in the first place. As of this commit, they all work fine with dinputto8.

* Prevent the enumeration functions from presenting DInput3 games with oversize structures

Games may be setting themselves for a bad time if they e.g. memcpy the structure
somewhere else and assume a maximum size, while also trying to respect a dwSize
parameter they received. To prevent potential buffer overflows, it's safer to clamp
the reported size.

* Mirrored behaviour changes from EnumObjects in GetObjectInfo